requirements.txt
python-vlc
pyDes
requests
Pillow
requests
glfw
pyyaml
numpy
baidu-aip
pyaudio
PyOpenGL
psd-tool
urllib